requirements.txt
flask
torch
torchvision
segmentation-models-pytorch
opencv-python
numpy
pillow
